feat: start moving functions to dependency
